@@ -1,6 +1,7 @@
 package br.com.ufg.tcc.medicamentos.classificationatc;
 
 import br.com.ufg.tcc.medicamentos.classificationatc.level.ClassificacionAtc;
+import br.com.ufg.tcc.medicamentos.common.ResourceNotFoundException;
 import br.com.ufg.tcc.medicamentos.medicament.MedicamentEntity;
 import br.com.ufg.tcc.medicamentos.medicament.MedicamentRepository;
 import org.apache.poi.ss.usermodel.*;
@@ -26,13 +27,16 @@ public class ClassificationAtcService {
     @Autowired
     private MedicamentRepository medRepository;
 
+    @Autowired
+    private AwsClassificationAtcService awsClassificationAtcService;
+
     public List<ClassificacionAtc> getAll() {
         List<ClassificationAtcEntity> all = repository.findAll();
         List<ClassificacionAtc> classifications = new ArrayList<>();
         List<ClassificacionAtc> classificacionAtcs = new ArrayList<>();
 
         all.forEach(a -> {
-            classifications.add(new ClassificacionAtc(a.getId(), a.getCodeAtc(), a.getName(), a.getLevel(), a.getCodeAtcParent()));
+            classifications.add(convert(a));
         });
 
         List<ClassificacionAtc> levelOne = classifications.stream().filter(c -> c.getLevel() == 1).collect(Collectors.toList());
@@ -48,18 +52,16 @@ public class ClassificationAtcService {
 
                 twoLevels.forEach(two -> {
                     List<ClassificacionAtc> threeLevels = getLevelsChildren(two, classifications);
-                    if (!threeLevels.isEmpty()) {
-                        two.setChildren(threeLevels);
+                    two.setChildren(threeLevels);
 
-                        threeLevels.forEach(three -> {
-                            List<ClassificacionAtc> fourLevels = getLevelsChildren(three, classifications);
-                            if (!fourLevels.isEmpty()) {
-                                three.setChildren(fourLevels);
-                            }
+                    threeLevels.forEach(three -> {
+                        List<ClassificacionAtc> fourLevels = getLevelsChildren(three, classifications);
+                        three.setChildren(fourLevels);
+                        fourLevels.forEach(four -> {
+                            List<ClassificacionAtc> fiveLevels = getLevelsChildren(four, classifications);
+                            four.setChildren(fourLevels);
                         });
-                    }
-
-
+                    });
                 });
 
                 classificacionAtcs.add(oneLevel);
@@ -71,8 +73,16 @@ public class ClassificationAtcService {
     }
 
     private List<ClassificacionAtc> getLevelsChildren(ClassificacionAtc level, List<ClassificacionAtc> classifications) {
-        List<ClassificacionAtc> levelsChildren = classifications.stream().filter(a -> a.getCodeAtcParent().equals(level.getCodeAtc()) && level.getCodeAtc() != a.getCodeAtc()).collect(Collectors.toList());
-        return levelsChildren;
+        List<ClassificacionAtc> levelsChildren = classifications.stream().filter(a -> Objects.nonNull(a.getCodeAtcParent()) && a.getCodeAtcParent().equals(level.getCodeAtc()) && level.getCodeAtc() != a.getCodeAtc()).collect(Collectors.toList());
+        return levelsChildren.isEmpty() ? new ArrayList<>() : levelsChildren;
+    }
+
+    public void save(ClassificacionAtc classificacionAtc) {
+        repository.save(convert(classificacionAtc));
+    }
+
+    public void sendDataToAws() {
+        awsClassificationAtcService.sendClassifications(getAll());
     }
 
     // Jeito rápido de ler o arquivo, definir um arquivo padrão em que poderíamos realizar o upload para alimentar o sistema.
@@ -201,4 +211,64 @@ public class ClassificationAtcService {
         }
     }
 
+    private ClassificationAtcEntity convert(ClassificacionAtc classificacionAtc) {
+        ClassificationAtcEntity entity = new ClassificationAtcEntity(Objects.nonNull(classificacionAtc.getId()) ? classificacionAtc.getId() : UUID.randomUUID());
+        entity.setCodeAtc(classificacionAtc.getCodeAtc());
+        entity.setCodeAtcParent(classificacionAtc.getCodeAtcParent());
+        entity.setLevel(classificacionAtc.getLevel());
+        entity.setName(classificacionAtc.getName());
+        return entity;
+    }
+
+    private ClassificacionAtc convert(ClassificationAtcEntity entity) {
+        ClassificacionAtc classificacionAtc = new ClassificacionAtc();
+        classificacionAtc.setCodeAtc(entity.getCodeAtc());
+        classificacionAtc.setCodeAtcParent(entity.getCodeAtcParent());
+        classificacionAtc.setLevel(entity.getLevel());
+        classificacionAtc.setName(entity.getName());
+        classificacionAtc.setId(entity.getId());
+        return classificacionAtc;
+    }
+
+    private List<ClassificacionAtc> convert(List<ClassificationAtcEntity> entitys) {
+        List<ClassificacionAtc> classifications = new ArrayList<>();
+        entitys.forEach(e -> classifications.add(convert(e)));
+        return classifications;
+    }
+
+    public ClassificacionAtc findById(UUID uuid) {
+        ClassificationAtcEntity entity = Optional.ofNullable(repository.findById(uuid).orElseThrow(RuntimeException::new)).get();
+        return convert(entity);
+    }
+
+    private void setChildren(ClassificacionAtc classificationAtc) {
+        List<ClassificationAtcEntity> children = repository.findByCodeAtcParent(classificationAtc.getCodeAtc());
+
+        if(Objects.nonNull(children)) {
+            classificationAtc.setChildren(convert(children));
+
+            classificationAtc.getChildren().forEach(c -> {
+                setChildren(c);
+            });
+        } else {
+            classificationAtc.setChildren(new ArrayList<>());
+        }
+    }
+
+    public ClassificacionAtc findByCode(String codeAtc) {
+        ClassificationAtcEntity entityOpt = repository.findByCode(codeAtc).orElseThrow(ResourceNotFoundException::new);
+        ClassificacionAtc classificationAtc = convert(entityOpt);
+
+        setChildren(classificationAtc);
+
+        return classificationAtc;
+    }
+
+    public void delete(UUID uuid) {
+        repository.deleteById(uuid);
+    }
+
+    public void update(ClassificacionAtc classificacionAtc) {
+        repository.update(classificacionAtc.getId(), classificacionAtc.getName());
+    }
 }
