@@ -1,15 +1,16 @@
 package br.com.ufg.tcc.medicamentos.classificationatc;
 
 import br.com.ufg.tcc.medicamentos.classificationatc.level.ClassificacionAtc;
+import br.com.ufg.tcc.medicamentos.common.ResponseApi;
 import io.swagger.annotations.Api;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 import springfox.documentation.annotations.ApiIgnore;
 
 import java.util.List;
+import java.util.UUID;
 
 @RestController
 @RequestMapping("/atc")
@@ -20,13 +21,41 @@ public class ClassificationAtcController {
     private ClassificationAtcService service;
 
     @GetMapping(value = "/test")
-    public String getAtcTeste() {
-        return "TESTE ENDPOINT ATC";
+    public ResponseEntity.BodyBuilder getTest() {
+        return ResponseEntity.ok();
     }
 
     @GetMapping(produces = "application/json")
-    public List<ClassificacionAtc> getAll() {
+    public List<ClassificacionAtc> findAll() {
         return service.getAll();
+    }
+
+    @GetMapping(path = "/{uuid}", produces = "application/json")
+    public ClassificacionAtc findById(@PathVariable UUID uuid) {
+        return service.findById(uuid);
+    }
+
+    @GetMapping(path = "/code/{codeAtc}", produces = "application/json")
+    public ClassificacionAtc findByCodeAtc(@PathVariable String codeAtc) {
+        return service.findByCode(codeAtc);
+    }
+
+    @PostMapping(consumes =  "application/json", produces = "application/json")
+    public ResponseEntity<ResponseApi> createClassificationAtc(@RequestBody ClassificacionAtc classificacionAtc) {
+        service.save(classificacionAtc);
+        return ResponseEntity.ok(ResponseApi.builder().message("Classificação ATC cadastrada com sucesso.").status(HttpStatus.OK).build());
+    }
+
+    @PutMapping(consumes =  "application/json", produces = "application/json")
+    public ResponseEntity<ResponseApi> updateClassificationAtc(@RequestBody ClassificacionAtc classificacionAtc) {
+        service.update(classificacionAtc);
+        return ResponseEntity.ok(ResponseApi.builder().message("Classificação ATC alterada com sucesso.").status(HttpStatus.OK).build());
+    }
+
+    @DeleteMapping(path = "/{uuid}", produces = "application/json")
+    public ResponseEntity<ResponseApi> deleteMedicament(@PathVariable UUID uuid) {
+        service.delete(uuid);
+        return ResponseEntity.ok(ResponseApi.builder().message("Classificação ATC deletado com sucesso.").status(HttpStatus.OK).build());
     }
 
     @ApiIgnore
